prepare build, rename namespaces, create badge
