+++ README.md
@@ -1,4 +1,5 @@
 [![Build Status](https://travis-ci.org/KadekM/scrawler.svg?branch=dev)](https://travis-ci.org/KadekM/scrawler)
+[![Maven Central](https://img.shields.io/maven-central/v/com.marekkadek/scrawler_2.11.svg)](https://maven-badges.herokuapp.com/maven-central/com.marekkadek/scrawler_2.11)
 
 ## SCRAWLER
