+++ modules/scraper/src/main/scala/com/marekkadek/scraper/browser.scala
@@ -1,6 +1,6 @@
-package com.kadekm.scraper
+package com.marekkadek.scraper
 
-import com.kadekm.scraper.proxy.ProxySettings
+import proxy.ProxySettings
 import fs2.util._
 
 abstract class Browser[F[_]](implicit FI: Effect[F]) {
