+++ modules/scraper/src/main/scala/com/marekkadek/scraper/dom.scala
@@ -1,4 +1,4 @@
-package com.kadekm.scraper
+package com.marekkadek.scraper
 
 /**
   * Represents HTML Document
