+++ modules/scraper/src/main/scala/com/marekkadek/scraper/extractors.scala
@@ -1,4 +1,4 @@
-package com.kadekm.scraper
+package com.marekkadek.scraper
 
 object extractors {
