+++ modules/scraper/src/main/scala/com/marekkadek/scraper/htmlunit/browser.scala
@@ -1,8 +1,7 @@
-package com.kadekm.scraper.htmlunit
+package com.marekkadek.scraper.htmlunit
 
-import com.kadekm.scraper._
+import com.marekkadek.scraper.Browser
 import fs2.util._
-import fs2._
 
 // Abstract until implemented
 abstract sealed class HtmlUnitBrowser[F[_]](implicit FI: Effect[F]) extends Browser[F] {}
