+++ modules/scraper/src/main/scala/com/marekkadek/scraper/package.scala
@@ -1,3 +1,3 @@
-package com.kadekm
+package com.marekkadek
 
 package object scraper {}
