+++ modules/scraper/src/main/scala/com/marekkadek/scraper/proxy.scala
@@ -1,4 +1,4 @@
-package com.kadekm.scraper
+package com.marekkadek.scraper
 
 import java.net._
