--- modules/scrawler/src/main/scala/com/kadekm/scrawler/package.scala
@@ -1,3 +1,0 @@
-package com.kadekm
-
-package object scrawler {}
