+++ modules/scrawler/src/test/scala/com/marekkadek/scrawler/Utils.scala
@@ -1,4 +1,4 @@
-package com.kadekm.scraper
+package com.marekkadek.scrawler
 
 import java.io.InputStream
